[COOK-2729] - qualify conditional w/ URI's regex
Use URI's ABS_URI regular expression to check if the URI is in fact
valid before checking if it has http or https scheme.

This catches the case of source being passed as "C:\\path\\to\file"
which may occur if someone uses the win_friendly_path in this very
module.

Or the case where someone really likes backslashes.