Remove vm_hostname:false from .travis.yml - Github Issue #189
This was effectively setting the hostname to false for all guests. Which we believe
was helping a bug where kitchen-vagrant/vagrant would fail to set the hostname if it was
over a certain length. Although, I had intended this to make kitchen-vagrant not set a
hostname at all, but apparently to that we needed this value to be nil.

That's the new default in the new-driver-model branch of kitchen-vagrant, which
we should see merged into master and/or windows-guest-support soon.