Update testing stack with the latest test-kitchen changes.
A new parameter was added to the driver config so that the correct
communicator is setup. Technically transport is not needed anymore
since it is discovered with the name, but I left it in there to be
explicit.

I bumped the gems to the latest working betas. Kitchen-vagrant
0.17.0.beta3 requires the vagrant-winrm plugin and will display
a message stating that you need to install it if you have not
already. I think this should suffice as a warning, but I could add
a special testing section to the README if necessary.

This has been tested with all 3 suites.