Add a root key attribute for the auto_run resource
I recently found myself working on a cookbook for an app that only installs on
a per-user basis, putting all its registry entries under the
`HKEY_CURRENT_USER` root key instead of `HKEY_LOCAL_MACHINE`.

This change adds a `root` property to the `windows_auto_run` resource,
maintaining the current behavior as a default and allowing support for
user-level applications as well.

Signed-off-by: Jonathan Hartman <j@hartman.io>