Create feature_cached.rb
This is a cached version of feature_dism. Dism is ran once and a copy is stored in memory and reused for the life time of chef-client.
The cache is cleared if pages are installed or removed via this same provisioner.

For now to use set node.default['windows']['feature_provider'] = 'cached'.
Perhaps once it's proven it can be added to feature.rb as a normal option.

I've tested this with the provided test kitchen resources as well as with our own cookbooks.