A cached feature provider for fast feature checks
A cached version for feature_dism.rb which enables lightening fast checks for installed windows packages. Very useful when you need to check many packages. As noted in some previous commits to feature_dism this process can be rather slow.

The only draw back to caching is that chef will be unaware of packages installed by other external processes during the converge which is why I have added this a separate provider for now.

Since the feature dsl checks node['windows']['feature_provider'] for the provider name the cached provider can be enabled via an attribute.