Improve version helper constant definition
Former constant definition was protected using the  method.
This method would return  if a module or a class exist with the same
name is found.

This commit improve the check to avoid multiple definition but also allow
definition of the constant if a module or class with the same name exist.

Signed-off-by: Baptiste Courtois <b.courtois@criteo.com>