Backport the windows_task resource from Chef 13.7 to 13.0-13.6
This is really not something I'd like to do in the future, but we need to get a stable release of the windows_task resource out to our existing Chef 13 windows base.  This version of the resource comes from the pre-Chef 13.7 release and will be used on Chef 13.0 - 13.6 so as to not impact Chef 12 users accustomed to the existing cookbook resource.

Signed-off-by: Tim Smith <tsmith@chef.io>