Default to dism for windows_feature instead of checking the path for dism
We're doing some unnecessary file IO here every time we run the resource instead of just doing the sensible thing and defaulting to dism. A lot of this was to support Windows releases that didn't have DISM, but we don't support those at this point so we should just default to dism unless someone specifies PowerShell.

Signed-off-by: Tim Smith <tsmith@chef.io>