Properly parse features into arrays
This took a bit longer to figure out than I care to admit. ["feature1, feature2"] is easy to miss. Turns out custom resources turn any name property into a string by joining arrays with commas. We then turned this into an array, but it was an array with a single item and that item wasn't actually a feature. So now we properly resplit the name property into an array so we can lookup each feature and figure out what to do.

Signed-off-by: Tim Smith <tsmith@chef.io>