Allow feature names to be case insensitive
The CLIs aren't actually case sensitive, but we asked for available features and only allowed in feature names that matched. Since we're doing a basic comparison of the array of desired features against what's available the case needs to match. The solution is to downcase the desired and available features and always work in lowercase. I added a test for both dism and powershell that uses the lowercase values.

Signed-off-by: Tim Smith <tsmith@chef.io>