Restore support for Windows 2008 R2 in windows_feature_dism
We technically don't support Windows 2008 R2 and it's EOL with Microsoft as well, but we can make this work with a bit of work. When we're on w2k8r2 just don't downcase. The user will have to provide a properly cased feature, but that was always the case on this version of Windows. It's a decade old.

Signed-off-by: Tim Smith <tsmith@chef.io>