Fixes a hang with font installs in some cases
Steps to repeat the bug:

* install a font from file "first_name.ttf" but with internal font name (i.e. in the font metadata within the file) of "my-font"
* install a second font from file "second_name.ttf" but with (same as above) internal font name "my-font"

The first font installs fine, the second hangs.

This occurs because windows see the second install as an overwrite of the first, and pops up a confirmation dialog box asking whether to override or cancel. Of course, running within the context of chef, this is not visible (but it can be replicated by running the same code from `install_font` in a ruby shell).

The fix here is to (hopefully) use the same mechanism as windows to determine font "equality" - i.e. metadata, rather than file name.

The downside is that candidate fonts to install must be downloaded first if they are remote, which can slow down provisioning (although not much if everything is running in the same data centre).

Signed-off-by: Jon Burgess <jkburges@gmail.com>